+++ packages/dds/merge-tree/src/test/mergeTree.insert.deltaCallback.spec.ts
@@ -6,7 +6,7 @@
 import { strict as assert } from "assert";
 import { LocalClientId, UnassignedSequenceNumber, UniversalSequenceNumber } from "../constants";
 import { MergeTreeMaintenanceType } from "../mergeTreeDeltaCallback";
-import { MergeTree } from "../mergeTreeInt";
+import { MergeTree } from "../mergeTree";
 import {
     MergeTreeDeltaType,
     ReferenceType,
