+++ packages/dds/merge-tree/src/test/mergeTree.insertingWalk.spec.ts
@@ -9,13 +9,13 @@ import { strict as assert } from "assert";
 import {
     IMergeBlock,
     MaxNodesInBlock,
-} from "../mergeTree";
+} from "../mergeTreeNodes";
 import {
     MergeTreeTextHelper,
     TextSegment,
 } from "../textSegment";
 import { LocalClientId, UnassignedSequenceNumber, UniversalSequenceNumber } from "../constants";
-import { MergeTree } from "../mergeTreeInt";
+import { MergeTree } from "../mergeTree";
 import { insertText, nodeOrdinalsHaveIntegrity } from "./testUtils";
 
 interface ITestTreeFactory {
