+++ packages/dds/merge-tree/src/test/text.ts
@@ -6,8 +6,8 @@
 import {
     ISegment,
     Marker,
-} from "../mergeTree";
-import { MergeTree } from "../mergeTreeInt";
+} from "../mergeTreeNodes";
+import { MergeTree } from "../mergeTree";
 import { ReferenceType } from "../ops";
 import { reservedTileLabelsKey } from "../referencePositions";
 import { TextSegment } from "../textSegment";
